code/TaskMan/requirements.txt
dataclasses==0.6
psutil==6.0.0
pyxdg==0.28
